+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Amélioration de la navigation lors de la modification d'un document (onglet, champs de données externe) #136
 - Modification des noms de fichier retour du Tdt pour actes #151
 - Envoi du nom original du fichier actes au parapheur #133
+- Ajout d'un bouton pour télécharger tous les fichiers d'un champs "fichier multiple" en même temps #185
 
 ## Ajouts